Extract pipeline steps from CLI
